[refactor] Move model utility functions to model_utils module
